feat(kafka): Allow custom partitioner for publisher
This commit introduces the  option, which allows a custom  to be set for the Kafka publisher.

By providing this option, users can define their own message partitioning strategy instead of relying on the default. This is useful for scenarios like key-based partitioning, which ensures messages with the same key are delivered to the same partition, thus guaranteeing order.